pyproject.toml
[project]
name = "curaitor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "arxiv==2.2.0",
    "asyncio>=4.0.0",
    "crawler==0.0.2",
    "dataclasses>=0.8",
    "datetime>=5.5",
    "extractor==0.5",
    "faiss-cpu==1.12.0",
    "google-adk>=1.14.1",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain==0.3.27",
    "langchain-google-genai==2.1.10",
    "langchain-ollama==0.3.8",
    "langchain-openai==0.3.33",
    "litellm>=1.77.1",
    "logging>=0.4.9.6",
    "mcp>=1.14.0",
    "npx>=0.1.6",
    "numpy==2.3.3",
    "pdfminer==20191125",
    "pylint>=3.3.8",
    "pypdf==6.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "requests==2.32.5",
    "scrapy==2.13.3",
    "sentence-transformers==5.1.0",
    "tiktoken==0.11.0",
    "typing>=3.10.0.0",
]
